Show messages in the progress bar
To do this we need to inherit from the pbar and use a few of it's internal details to our advantage. Knowing that the progressbar did not change in the last 5 years we can count on it continuing it's fast pace of development and we can consider it's API stable.