Switch to jeweler instead of Hoe
We need to exclude a globbed pattern since it's not desirable for everyone to be downloading the whole test corpus every time an update comes out. The packaged gem size has been squashed to 80K-ish from 4meg-ish, which is awesome and will let everyone update faster.
We also want to be on Travis and somwhere in the "Kramer vs. Kramer"-like affair between rubygems, hoe and the travis team everything has gone sour. So solve a bulk of issues with that. We also specify our deps once and only once.