Implicit block capture is no longer supported
Just doing Proc.new without a block is no longer supported. It used to be a hack used
for avoiding passing the `blk` block argument - which used to work faster on most Ruby versions.

Instead add usual block captures and use `#to_proc` or & to pass them downstream to the callees.